Move composer API creation
