+++ apps/meteor/client/startup/afterLogoutCleanUp/purgeAllDrafts.ts
@@ -1,7 +1,7 @@
 import { Meteor } from 'meteor/meteor';
 
+import { purgeAllDrafts } from '../../../app/ui-message/client/messageBox/messageBox';
 import { callbacks } from '../../../lib/callbacks';
-import { purgeAllDrafts } from '../../lib/chats/composer';
 
 Meteor.startup(() => {
 	callbacks.add('afterLogoutCleanUp', purgeAllDrafts, callbacks.priority.MEDIUM, 'chatMessages-after-logout-cleanup');
